requirements.txt
python-docx
reportlab
pdfminer.six
markdown
pypandoc
